+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/permission/ui/MiniAppCustomPermissionAdapter.kt
@@ -93,6 +93,7 @@ internal class MiniAppCustomPermissionAdapter :
             MiniAppCustomPermissionType.PROFILE_PHOTO -> "Profile Photo"
             MiniAppCustomPermissionType.CONTACT_LIST -> "Contact List"
             MiniAppCustomPermissionType.ACCESS_TOKEN -> "Access Token"
+            MiniAppCustomPermissionType.SEND_MESSAGE -> "Send Message"
             MiniAppCustomPermissionType.LOCATION -> "Device Location"
             else -> "Unknown"
         }